Moved split Number to own Util file
